Don't use d as a spell letter
.. as it conflicts with d-for-drop in the spellbook UI. d is still used
as a spell letter if there is no source item, however (i.e. if the spell
list is coming from a monster), as that doesn't present conflicts.

This commit also reimplements the storage for spell-letter pairings, as
the previous implementation had two different structures and did not
handle missing letters.