Add saving mute lists via a cookie
Still not really sure if this is the right way to go, but this allows
saving (per-channel, player-managed) mute lists via a cookie in the
player's browser. This probably won't yet work with a console player
spectating themselves (they should be able to mute spectators, but the
list won't save).